+++ apps/astarte_appengine_api/lib/astarte_appengine_api/application.ex
@@ -19,6 +19,8 @@ defmodule Astarte.AppEngine.API.Application do
   use Application
   require Logger
 
+  @app_version Mix.Project.config()[:version]
+
   # See https://hexdocs.pm/elixir/Application.html
   # for more information on OTP Applications
   def start(_type, _args) do
@@ -32,7 +34,7 @@ defmodule Astarte.AppEngine.API.Application do
       {&:logger_filters.domain/2, {:stop, :equal, [:progress]}}
     )
 
-    Logger.info("Starting application.", tag: "appengine_api_start")
+    Logger.info("Starting application v#{@app_version}.", tag: "appengine_api_start")
 
     DataAccessConfig.validate!()
     RPCConfig.validate!()